Bring changes up to speed with mockup
