+++ frontend/src/App.scss
@@ -115,6 +115,10 @@ body {
   color: $text-ink
 }
 
+.base-medium {
+  color: $base-medium;
+}
+
 .smart-hub-bg-blue-primary {
   background-color: $ttahub-blue;
 }
