+++ frontend/src/colors.scss
@@ -21,6 +21,7 @@ $ttahub-orange: #e29f4d;
 $ttahub-orange-light: #fff1e0;
 $base-darkest: #1b1b1b;
 $base-dark: #565c65;
+$base-medium: #71767A;
 $base-light: #a9aeb1;
 $base-lighter: #dfe1e2;
 $base-lightest: #f0f0f0;
